# Change ArrayBlock to use arrays internally