--- a/presto-spi/src/main/java/com/facebook/presto/spi/block/ArrayBlock.java
+++ b/presto-spi/src/main/java/com/facebook/presto/spi/block/ArrayBlock.java
@@ -13,9 +13,10 @@
  */
 package com.facebook.presto.spi.block;
 
-import io.airlift.slice.Slice;
 import org.openjdk.jol.info.ClassLayout;
 
+import static com.facebook.presto.spi.block.BlockUtil.intSaturatedCast;
+import static io.airlift.slice.SizeOf.sizeOf;
 import static java.util.Objects.requireNonNull;
 
 public class ArrayBlock
@@ -23,35 +24,66 @@ public class ArrayBlock
 {
     private static final int INSTANCE_SIZE = ClassLayout.parseClass(ArrayBlock.class).instanceSize();
 
+    private final int arrayOffset;
+    private final int positionCount;
+    private final boolean[] valueIsNull;
     private final Block values;
-    private final Slice offsets;
-    private final int offsetBase;
-    private final Slice valueIsNull;
+    private final int[] offsets;
 
-    public ArrayBlock(Block values, Slice offsets, int offsetBase, Slice valueIsNull)
+    private final int sizeInBytes;
+    private final int retainedSizeInBytes;
+
+    public ArrayBlock(int positionCount, boolean[] valueIsNull, int[] offsets, Block values)
     {
+        this(0, positionCount, valueIsNull, offsets, values);
+    }
+
+    ArrayBlock(int arrayOffset, int positionCount, boolean[] valueIsNull, int[] offsets, Block values)
+    {
+        if (arrayOffset < 0) {
+            throw new IllegalArgumentException("arrayOffset is negative");
+        }
+        this.arrayOffset = arrayOffset;
+
+        if (positionCount < 0) {
+            throw new IllegalArgumentException("positionCount is negative");
+        }
+        this.positionCount = positionCount;
+
+        requireNonNull(valueIsNull, "valueIsNull is null");
+        if (valueIsNull.length - arrayOffset < positionCount) {
+            throw new IllegalArgumentException("isNull length is less than positionCount");
+        }
+        this.valueIsNull = valueIsNull;
+
+        requireNonNull(offsets, "offsets is null");
+        if (offsets.length - arrayOffset < positionCount + 1) {
+            throw new IllegalArgumentException("offsets length is less than positionCount");
+        }
+        this.offsets = offsets;
+
         this.values = requireNonNull(values);
-        this.offsets = requireNonNull(offsets);
-        this.offsetBase = offsetBase;
-        this.valueIsNull = requireNonNull(valueIsNull);
+
+        sizeInBytes = values.getSizeInBytes() + ((Integer.BYTES + Byte.BYTES) * this.positionCount);
+        retainedSizeInBytes = intSaturatedCast(INSTANCE_SIZE + values.getRetainedSizeInBytes() + sizeOf(offsets) + sizeOf(valueIsNull));
     }
 
     @Override
     public int getPositionCount()
     {
-        return valueIsNull.length();
+        return positionCount;
     }
 
     @Override
     public int getSizeInBytes()
     {
-        return getValues().getSizeInBytes() + offsets.length() + valueIsNull.length();
+        return sizeInBytes;
     }
 
     @Override
     public int getRetainedSizeInBytes()
     {
-        return INSTANCE_SIZE + values.getRetainedSizeInBytes() + offsets.getRetainedSize() + valueIsNull.getRetainedSize();
+        return retainedSizeInBytes;
     }
 
     @Override
@@ -61,7 +93,7 @@ public class ArrayBlock
     }
 
     @Override
-    protected Slice getOffsets()
+    protected int[] getOffsets()
     {
         return offsets;
     }
@@ -69,11 +101,11 @@ public class ArrayBlock
     @Override
     protected int getOffsetBase()
     {
-        return offsetBase;
+        return arrayOffset;
     }
 
     @Override
-    protected Slice getValueIsNull()
+    protected boolean[] getValueIsNull()
     {
         return valueIsNull;
     }