Change ArrayBlock offsets to include an extra 0 entry in the first position.
With this change the length and offset functions do not need an if statement.