@haozhun 2016-11-16
Travis failure appears to be relevant