@haozhun 2016-11-21
```
Before
Benchmark   (array/mapSize)      (name)  Mode  Cnt      Score     Error  Units
arraySubscript            1   fix-width  avgt   20     17.743 ±   0.452  ns/op
arraySubscript            1   var-width  avgt   20     47.122 ±   1.594  ns/op
arraySubscript            1  dictionary  avgt   20    190.536 ±   8.326  ns/op
arraySubscript            1       array  avgt   20    101.283 ±   2.736  ns/op
arraySubscript           13   fix-width  avgt   20    280.263 ±   6.887  ns/op
arraySubscript           13   var-width  avgt   20    720.084 ±  20.839  ns/op
arraySubscript           13  dictionary  avgt   20  13341.996 ± 318.443  ns/op
arraySubscript           13       array  avgt   20   1451.829 ±  28.955  ns/op
mapSubscript              1   fix-width  avgt   20     47.678 ±   1.440  ns/op
mapSubscript              1   var-width  avgt   20     97.814 ±   2.253  ns/op
mapSubscript              1  dictionary  avgt   20     81.814 ±   1.320  ns/op
mapSubscript             13   fix-width  avgt   20   2477.750 ±  60.531  ns/op
mapSubscript             13   var-width  avgt   20   3340.492 ±  96.619  ns/op
mapSubscript             13  dictionary  avgt   20   3322.502 ±  75.192  ns/op

After
Benchmark   (array/mapSize)      (name)  Mode  Cnt      Score     Error  Units
arraySubscript            1   fix-width  avgt   20     14.766 ±   0.649  ns/op
arraySubscript            1   var-width  avgt   20     40.475 ±   1.245  ns/op
arraySubscript            1  dictionary  avgt   20    162.354 ±   8.634  ns/op
arraySubscript            1       array  avgt   20     60.295 ±   1.495  ns/op
arraySubscript           13   fix-width  avgt   20    210.913 ±   1.944  ns/op
arraySubscript           13   var-width  avgt   20    638.209 ±   8.453  ns/op
arraySubscript           13  dictionary  avgt   20  11678.850 ± 113.359  ns/op
arraySubscript           13       array  avgt   20    909.282 ±  13.197  ns/op
mapSubscript              1   fix-width  avgt   20     44.117 ±   1.317  ns/op
mapSubscript              1   var-width  avgt   20     94.888 ±   2.639  ns/op
mapSubscript              1  dictionary  avgt   20    101.197 ±   2.391  ns/op
mapSubscript             13   fix-width  avgt   20   2318.569 ±  50.811  ns/op
mapSubscript             13   var-width  avgt   20   3032.081 ± 109.539  ns/op
mapSubscript             13  dictionary  avgt   20   3221.476 ± 172.099  ns/op
```