@haozhun 2016-11-22
Looks good.

The benchmark item marked `*****` is 15% worse. I'm not sure whether it matters enough. What do you think @cberner?