@cberner 2016-11-23
Ya, seems fine. We're going to have to do some performance optimization for map subscript at some point, anyway. However, the ones that probably need it the most are large maps, rather than small ones